fix:  Remove invalid HeadCalendar export field
